Correct detection of the fixed fonts glyph offset size.
The problem of the original test is, that there is no guarantee,
that array of 16-bit integers is 32-bit aligned. This means,
that on architectures without support for unaligned memory access
the test results in bogus results in unpredictable cases.
The test did not work on big endian architectures at all.
The long pointing to the index 0 overlaps in MSB bits with
index 0 of the short ints, which holds zero offset typically,
so the result has marked font as long offset type always.

Signed-off-by: Pavel Pisa <pisa@cmp.felk.cvut.cz>