MW: fixes of type definitions to build on 64-bit systems.
The long type is 64-bit on x86_64 systems. Use correctly sized
types for memory operations. The inttypes.h is used as base for
size describing types. The use of inttypes.h should not be problem
for build on any system already, because this header file is part
of C standard for 10 years already.

Signed-off-by: Pavel Pisa <pisa@cmp.felk.cvut.cz>